Cleaned up files. added navigation back to admin home after adding pet.
