+++ plentyofpets/lib/models/pet_model.dart
@@ -1,13 +1,21 @@
 
 
+//Pet class for use in creating list of pet objects from query Stream
+
 class Pets {
 
   final String type;
   final String availability;
-  //final String disposition;
+  final String disposition;
   final String breed;
   final String name;
-  //final String description;
+  final String description;
 
-  Pets({required this.name,required this.availability,required this.breed,required this.type});
+  Pets({
+    required this.type,
+    required this.availability,
+    required this.disposition,
+    required this.breed,
+    required this.name,
+    required this.description});
 }
